refactor: move stubs and mocks to a file
