+++ src/__tests__/use-case/auth-group/db-create-auth-group.spec.ts
@@ -1,8 +1,6 @@
-import { AuthGroupProperties, AuthGroup } from '@/entity/auth-group'
-import { uuid } from '@/entity/utils'
 import { DbCreateAuthGroup } from '@/use-case/auth-group/db-create-auth-group'
 import { CreateAuthGroupRepository } from '@/use-case/auth-group/protocols/create-auth-group-repository'
-import faker from 'faker'
+import { makeCreateAuthGroupRepositoryStub, mockedAuthGroup, mockedAuthGroupProperties } from '../stubs/auth-group'
 
 describe('DbCreateAuthGroup', () => {
   it('should call CreateAuthGroupRepository with correct values', async () => {
@@ -40,37 +38,4 @@ function makeSut (): SutTypes {
     sut,
     createAuthGroupStub
   }
-}
-
-function makeCreateAuthGroupRepositoryStub (): CreateAuthGroupRepository {
-  class CreateAuthGroupRepositoryStub implements CreateAuthGroupRepository {
-    async create (authGroup: AuthGroupProperties): Promise<AuthGroup> {
-      return mockedAuthGroup
-    }
-  }
-  return new CreateAuthGroupRepositoryStub()
-}
-
-const mockedAuthGroup = mockAuthGroup()
-function mockAuthGroup (): AuthGroup {
-  const activity = {
-    name: faker.name.title(),
-    permissions: [
-      faker.lorem.word(5),
-      faker.lorem.word(5),
-      faker.lorem.word(5)
-    ]
-  }
-  return {
-    id: <uuid>faker.datatype.uuid(),
-    title: faker.name.title(),
-    activities: [activity]
-  }
-}
-
-const mockedAuthGroupProperties = mockAuthGroupProperties()
-function mockAuthGroupProperties (): AuthGroupProperties {
-  const data = Object.assign({}, mockedAuthGroup)
-  Reflect.deleteProperty(data, 'id')
-  return data
 }
