+++ src/__tests__/use-case/stubs/auth-group.ts
@@ -1,0 +1,15 @@
+import { AuthGroupProperties, AuthGroup } from '@/entity/auth-group'
+import { CreateAuthGroupRepository } from '@/use-case/auth-group/protocols/create-auth-group-repository'
+import { mockAuthGroup, mockAuthGroupProperties } from '@/__tests__/entity/mock/auth-group'
+
+export const mockedAuthGroup = mockAuthGroup()
+export const mockedAuthGroupProperties = mockAuthGroupProperties()
+
+export function makeCreateAuthGroupRepositoryStub (): CreateAuthGroupRepository {
+  class CreateAuthGroupRepositoryStub implements CreateAuthGroupRepository {
+    async create (authGroup: AuthGroupProperties): Promise<AuthGroup> {
+      return mockedAuthGroup
+    }
+  }
+  return new CreateAuthGroupRepositoryStub()
+}
